+++ xl-evaluator/test/src/com/tjclp/xl/formula/FutureFunctionRegistrySpec.scala
@@ -19,6 +19,7 @@ class FutureFunctionRegistrySpec extends FunSuite:
     assertEquals(
       prefixed,
       Set(
+        "ANCHORARRAY",
         "FILTER",
         "IFNA",
         "IFS",